The framebuffer driver. Vastly rewritten for the new FB-Api. Also contains stuff that was lying in my shelf for a year but which I never added.
With this driver all possible graphics resolutions are possible, also
all possible modes can be put on screen.

You guessed it, it's not completely tested (only on eNX for now).

Note that this can not be compiled with 2.4 (the #ifdef would have
practically split the code in two big halves ;) ).

Hardware acceleration (yes, there is some...) is currently missing.